+++ src/ViewInstance.js
@@ -6,7 +6,10 @@ import {
   FormattedMessage,
   injectIntl,
 } from 'react-intl';
-import { flowRight } from 'lodash';
+import {
+  flowRight,
+  isEmpty,
+} from 'lodash';
 
 import {
   AppIcon,
@@ -24,6 +27,7 @@ import {
   HasCommand,
   collapseAllSections,
   expandAllSections,
+  ConfirmationModal,
 } from '@folio/stripes/components';
 
 import ViewHoldingsRecord from './ViewHoldingsRecord';
@@ -164,6 +168,13 @@ class ViewInstance extends React.Component {
       tenant: '!{tenantId}',
       throwErrors: false,
     },
+    shareInstance: {
+      type: 'okapi',
+      path: 'consortia/!{consortiumId}/sharing/instances',
+      accumulate: true,
+      throwErrors: false,
+      tenant: '!{centralTenantId}',
+    },
   });
 
   constructor(props) {
@@ -178,6 +189,7 @@ class ViewInstance extends React.Component {
       isItemsMovement: false,
       isImportRecordModalOpened: false,
       isCopyrightModalOpened: false,
+      isShareLocalInstanceModalOpen: false,
       isNewOrderModalOpen: false,
       afterCreate: false,
       instancesQuickExportInProgress: false,
@@ -410,6 +422,35 @@ class ViewInstance extends React.Component {
     this.setState({ isImportRecordModalOpened: false });
   }
 
+  handleShareLocalInstance = (instance = {}) => {
+    const centralTenantId = this.props.centralTenantId;
+    const sourceTenantId = this.props.okapi.tenant;
+    const instanceTitle = instance.title;
+    const instanceIdentifier = instance.id;
+
+    this.props.mutator.shareInstance.POST({
+      sourceTenantId,
+      instanceIdentifier,
+      targetTenantId: centralTenantId,
+    })
+      .then(async () => {
+        await this.props.refetchInstance();
+        this.calloutRef.current.sendCallout({
+          type: 'success',
+          message: <FormattedMessage id="ui-inventory.shareLocalInstance.toast.successful" values={{ instanceTitle }} />,
+        });
+      })
+      .catch(() => {
+        this.calloutRef.current.sendCallout({
+          type: 'error',
+          message: <FormattedMessage id="ui-inventory.shareLocalInstance.toast.unsuccessful" values={{ instanceTitle }} />,
+        });
+      })
+      .finally(() => {
+        this.setState({ isShareLocalInstanceModalOpen: false });
+      });
+  }
+
   toggleCopyrightModal = () => {
     this.setState(prevState => ({ isCopyrightModalOpened: !prevState.isCopyrightModalOpened }));
   };
@@ -424,13 +465,14 @@ class ViewInstance extends React.Component {
 
   // Get all identifiers for all records
   getIdentifiers = (data) => {
-    const { identifierTypesById } = data;
-    const { ISBN, ISSN } = indentifierTypeNames;
     const selectedInstance = this.props?.selectedInstance;
 
-    if (!selectedInstance) {
+    if (!selectedInstance || isEmpty(data)) {
       return null;
     }
+
+    const { identifierTypesById } = data;
+    const { ISBN, ISSN } = indentifierTypeNames;
 
     // We can't make any meaningful assessment of which is
     // the best identifier to return, so just return the first
@@ -460,7 +502,6 @@ class ViewInstance extends React.Component {
       onCopy,
       stripes,
       intl,
-      openedFromBrowse,
       resources: {
         instanceRequests,
       },
@@ -469,10 +510,11 @@ class ViewInstance extends React.Component {
       marcRecord,
       titleLevelRequestsFeatureEnabled,
     } = this.state;
+    const source = instance?.source;
 
     const editBibRecordPerm = 'ui-quick-marc.quick-marc-editor.all';
     const editInstancePerm = 'ui-inventory.instance.edit';
-    const isSourceMARC = isMARCSource(instance?.source);
+    const isSourceMARC = isMARCSource(source);
     const canEditInstance = stripes.hasPerm(editInstancePerm);
     const canCreateInstance = stripes.hasPerm('ui-inventory.instance.create');
     const canCreateRequest = stripes.hasPerm('ui-requests.create');
@@ -488,12 +530,13 @@ class ViewInstance extends React.Component {
     const canViewMARCSource = stripes.hasPerm('ui-quick-marc.quick-marc-editor.view');
     const canViewInstance = stripes.hasPerm('ui-inventory.instance.view');
     const canViewSource = canViewMARCSource && canViewInstance;
+    const canShareLocalInstance = checkIfUserInMemberTenant(stripes) && !isShared && !isInstanceShadowCopy(source);
     const canCreateOrder = !checkIfUserInCentralTenant(stripes) && stripes.hasInterface('orders') && stripes.hasPerm('ui-inventory.instance.createOrder');
     const canReorder = stripes.hasPerm('ui-requests.reorderQueue');
     const numberOfRequests = instanceRequests.other?.totalRecords;
     const canReorderRequests = titleLevelRequestsFeatureEnabled && hasReorderPermissions && numberOfRequests && canReorder;
     const canViewRequests = !checkIfUserInCentralTenant(stripes) && !titleLevelRequestsFeatureEnabled;
-    const canCreateNewRequest = titleLevelRequestsFeatureEnabled && canCreateRequest;
+    const canCreateNewRequest = !checkIfUserInCentralTenant(stripes) && titleLevelRequestsFeatureEnabled && canCreateRequest;
     const identifier = this.getIdentifiers(data);
 
     const buildOnClickHandler = onClickHandler => {
@@ -511,7 +554,7 @@ class ViewInstance extends React.Component {
     const showInventoryMenuSection = (
       canEditInstance
       || canViewSource
-      || (!openedFromBrowse && (canMoveItems || canMoveHoldings))
+      || (canMoveItems || canMoveHoldings)
       || canUseSingleRecordImport
       || canCreateInstance
       || canCreateOrder
@@ -561,31 +604,27 @@ class ViewInstance extends React.Component {
                 disabled={!marcRecord}
               />
             )}
-            {!openedFromBrowse && (
-              <>
-                {canMoveItems && (
-                  <ActionItem
-                    id="move-instance-items"
-                    icon="transfer"
-                    messageId={`ui-inventory.moveItems.instance.actionMenu.${this.state.isItemsMovement ? 'disable' : 'enable'}`}
-                    onClickHandler={() => {
-                      onToggle();
-                      this.toggleItemsMovement();
-                    }}
-                  />
-                )}
-                {(canMoveItems || canMoveHoldings) && (
-                  <ActionItem
-                    id="move-instance"
-                    icon="arrow-right"
-                    messageId="ui-inventory.moveItems"
-                    onClickHandler={() => {
-                      onToggle();
-                      this.toggleFindInstancePlugin();
-                    }}
-                  />
-                )}
-              </>
+            {canMoveItems && (
+              <ActionItem
+                id="move-instance-items"
+                icon="transfer"
+                messageId={`ui-inventory.moveItems.instance.actionMenu.${this.state.isItemsMovement ? 'disable' : 'enable'}`}
+                onClickHandler={() => {
+                  onToggle();
+                  this.toggleItemsMovement();
+                }}
+              />
+            )}
+            {(canMoveItems || canMoveHoldings) && (
+              <ActionItem
+                id="move-instance"
+                icon="arrow-right"
+                messageId="ui-inventory.moveItems"
+                onClickHandler={() => {
+                  onToggle();
+                  this.toggleFindInstancePlugin();
+                }}
+              />
             )}
             {canUseSingleRecordImport && (
               <ActionItem
@@ -606,6 +645,17 @@ class ViewInstance extends React.Component {
                 onClickHandler={() => {
                   onToggle();
                   onCopy(instance);
+                }}
+              />
+            )}
+            {canShareLocalInstance && (
+              <ActionItem
+                id="share-local-instance"
+                icon="graph"
+                messageId="ui-inventory.shareLocalInstance"
+                onClickHandler={() => {
+                  onToggle();
+                  this.setState({ isShareLocalInstanceModalOpen: true });
                 }}
               />
             )}
@@ -648,7 +698,7 @@ class ViewInstance extends React.Component {
               )
             }
             <NewInstanceRequestButton
-              isTlrEnabled={!!titleLevelRequestsFeatureEnabled}
+              isTlrEnabled={canCreateNewRequest}
               instanceId={instance.id}
             />
           </MenuSection>
@@ -881,6 +931,15 @@ class ViewInstance extends React.Component {
               onCancel={this.toggleNewOrderModal}
             />
 
+            <ConfirmationModal
+              open={this.state.isShareLocalInstanceModalOpen}
+              heading={<FormattedMessage id="ui-inventory.shareLocalInstance.modal.header" />}
+              message={<FormattedMessage id="ui-inventory.shareLocalInstance.modal.message" values={{ instanceTitle: instance?.title }} />}
+              confirmLabel={<FormattedMessage id="ui-inventory.shareLocalInstance.modal.confirmButton" />}
+              onCancel={() => this.setState({ isShareLocalInstanceModalOpen: false })}
+              onConfirm={() => this.handleShareLocalInstance(instance)}
+            />
+
           </HasCommand>
         )}
       </DataContext.Consumer>
@@ -893,6 +952,8 @@ ViewInstance.propTypes = {
   canUseSingleRecordImport: PropTypes.bool,
   centralTenantPermissions: PropTypes.arrayOf(PropTypes.object).isRequired,
   selectedInstance:  PropTypes.object,
+  centralTenantId: PropTypes.string,
+  refetchInstance: PropTypes.func,
   goTo: PropTypes.func.isRequired,
   location: PropTypes.shape({
     pathname: PropTypes.string.isRequired,
@@ -928,10 +989,10 @@ ViewInstance.propTypes = {
       GET: PropTypes.func.isRequired,
       reset: PropTypes.func.isRequired,
     }).isRequired,
+    shareInstance: PropTypes.shape({ POST: PropTypes.func.isRequired }).isRequired,
   }),
   onClose: PropTypes.func,
   onCopy: PropTypes.func,
-  openedFromBrowse: PropTypes.bool,
   paneWidth: PropTypes.string.isRequired,
   resources: PropTypes.shape({
     allInstanceItems: PropTypes.object.isRequired,
@@ -952,6 +1013,7 @@ ViewInstance.propTypes = {
     locale: PropTypes.string.isRequired,
     logger: PropTypes.object.isRequired,
   }).isRequired,
+  okapi: PropTypes.object.isRequired,
   tagsEnabled: PropTypes.bool,
   updateLocation: PropTypes.func.isRequired,
 };