+++ src/components/HoldingsRecordFilters/HoldingsRecordFilters.test.js
@@ -22,6 +22,7 @@ jest.mock('../../facetUtils', () => ({
 
 const activeFilters = {
   [FACETS.SHARED]: ['shared1'],
+  [FACETS.HELD_BY]: ['HELD_BY1'],
   [FACETS.EFFECTIVE_LOCATION]: ['loc1'],
   [FACETS.HOLDINGS_PERMANENT_LOCATION]: ['loc2'],
   [FACETS.HOLDINGS_TYPE]: ['loc3'],
@@ -37,6 +38,7 @@ const resources = {
     resource: 'facets',
     records: [{
       'shared': { values: ['shared1'] },
+      'holdings.tenantId': { values: ['heldby1'] },
       'items.effectiveLocationId': { values: ['effectiveLocationId1'] },
       'holdings.permanentLocationId': { values: ['permanentLocationId1'] },
       'holdings.statisticalCodeIds': { values: ['statisticalCodeIds1'] },
@@ -59,6 +61,7 @@ const data = {
   modesOfIssuance: [],
   tagsRecords: [],
   natureOfContentTerms: [],
+  consortiaTenants: [],
   query: [],
   onFetchFacets: jest.fn(),
   parentResources: resources,
@@ -91,7 +94,7 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render shared, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const shared = document.querySelector('[id="accordion-toggle-button-shared"]');
+    const shared = screen.getByRole('button', { name: /Shared/i });
     userEvent.click(shared);
     const Clearselectedfilters = screen.getAllByRole('button');
     userEvent.click(Clearselectedfilters[1]);
@@ -100,12 +103,23 @@ describe('HoldingsRecordFilters', () => {
     });
   });
 
-  it('Should Render effectiveLocation, Clear selectedfilters buttons', async () => {
+  it('Should Render Held by, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const effectiveLocation = document.querySelector('[id="accordion-toggle-button-effectiveLocation"]');
-    userEvent.click(effectiveLocation);
+    const heldBy = screen.getByRole('button', { name: /held by/i });
+    userEvent.click(heldBy);
     const Clearselectedfilters = screen.getAllByRole('button');
     userEvent.click(Clearselectedfilters[3]);
+    await waitFor(() => {
+      expect(onClear).toBeCalledWith(FACETS.HELD_BY);
+    });
+  });
+
+  it('Should Render effectiveLocation, Clear selectedfilters buttons', async () => {
+    renderHoldingsRecordFilters();
+    const effectiveLocation = screen.getByRole('button', { name: /Effective location \(item\)/i });
+    userEvent.click(effectiveLocation);
+    const Clearselectedfilters = screen.getAllByRole('button');
+    userEvent.click(Clearselectedfilters[5]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.EFFECTIVE_LOCATION);
     });
@@ -113,10 +127,10 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsPermanentLocation, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsPermanentLocation = document.querySelector('[id="accordion-toggle-button-holdingsPermanentLocation"]');
+    const holdingsPermanentLocation = screen.getByRole('button', { name: /Holdings permanent location/i });
     userEvent.click(holdingsPermanentLocation);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[5]);
+    userEvent.click(Clearselectedfilters[7]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_PERMANENT_LOCATION);
     });
@@ -124,10 +138,10 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsType, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsType = document.querySelector('[id="accordion-toggle-button-holdingsType"]');
+    const holdingsType = screen.getByRole('button', { name: /Holdings type/i });
     userEvent.click(holdingsType);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[7]);
+    userEvent.click(Clearselectedfilters[9]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_TYPE);
     });
@@ -135,10 +149,10 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsDiscoverySuppress, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsDiscoverySuppress = document.querySelector('[id="accordion-toggle-button-holdingsDiscoverySuppress"]');
+    const holdingsDiscoverySuppress = screen.getByRole('button', { name: /Suppress from discovery/i });
     userEvent.click(holdingsDiscoverySuppress);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[9]);
+    userEvent.click(Clearselectedfilters[11]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_DISCOVERY_SUPPRESS);
     });
@@ -146,10 +160,10 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsStatisticalCodeIds, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsStatisticalCodeIds = document.querySelector('[id="accordion-toggle-button-holdingsStatisticalCodeIds"]');
+    const holdingsStatisticalCodeIds = screen.getByRole('button', { name: /Statistical code/i });
     userEvent.click(holdingsStatisticalCodeIds);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[11]);
+    userEvent.click(Clearselectedfilters[13]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_STATISTICAL_CODE_IDS);
     });
@@ -157,10 +171,10 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsCreatedDate, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsCreatedDate = document.querySelector('[id="accordion-toggle-button-holdingsCreatedDate"]');
+    const holdingsCreatedDate = screen.getByRole('button', { name: /Date created/i });
     userEvent.click(holdingsCreatedDate);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[13]);
+    userEvent.click(Clearselectedfilters[15]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_CREATED_DATE);
     });
@@ -168,20 +182,20 @@ describe('HoldingsRecordFilters', () => {
 
   it('Should Render holdingsUpdatedDate, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsUpdatedDate = document.querySelector('[id="accordion-toggle-button-holdingsUpdatedDate"]');
+    const holdingsUpdatedDate = screen.getByRole('button', { name: /Date updated/i });
     userEvent.click(holdingsUpdatedDate);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[20]);
+    userEvent.click(Clearselectedfilters[22]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_UPDATED_DATE);
     });
   });
   it('Should Render holdingsSource, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
-    const holdingsSource = document.querySelector('[id="accordion-toggle-button-holdingsSource"]');
+    const holdingsSource = screen.getByRole('button', { name: /Source/i });
     userEvent.click(holdingsSource);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[27]);
+    userEvent.click(Clearselectedfilters[29]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_SOURCE);
     });