+++ src/routes/ViewHoldingRoute.test.js
@@ -1,7 +1,10 @@
 import '../../test/jest/__mock__';
 
 import { MemoryRouter } from 'react-router-dom';
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import {
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
 
 import ViewHoldingRoute from './ViewHoldingRoute';